agent/requirements.txt
livekit-agents>=1.1.6
livekit-plugins-openai>=1.1.6
livekit-plugins-deepgram>=1.1.6
livekit-plugins-silero>=1.1.6
livekit-plugins-cartesia>=1.1.6
livekit-plugins-turn-detector>=1.1.6
python-dotenv
requests
chromadb>=0.4.0
numpy==1.26.4
